Separate the concept of labels from instances
We can't assume instances for an axis are also the labels.
For example, Arapey has two instances that don't only change
the `whgt` but also the `opsz` value.

As a quick and dirty solution I duplicated the generated
instances.json file, removed the opsz ones, and used that
as a list to deduct the tick labels from.

This is not ideal because instance steps don't have to be
equally distributed.

But for now it works :)